Split up giant member model
